Aperfeiçado a ideia de Ação. Ações devem ter uma duração fixa pré-determinada, ao registrar uma ação deverá ser informado em qual "server-tick" a ação estará concluida.

+++ 1stPrototype/Classes/Relacionamentos/EfeitoAcao.cs
@@ -10,6 +10,6 @@ namespace _1stPrototype.Classes.Relacionamentos
     public class EfeitoAcao
     {
         public enumEfeitoAcao efeito;
-        public double teor;
+        public float teor; // Teor total a ser aplicado.
     }
 }
